# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/blocks": {
            "get": {
                "description": "get latest n block information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Block Chain Information"
                ],
                "summary": "get latest n block information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.GetLatestNBlockRes"
                        }
                    },
                    "204": {
                        "description": "no content"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/blocks/{id}": {
            "get": {
                "description": "get block information by block number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Block Chain Information"
                ],
                "summary": "get block information by block number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.GetBlockByNumRes"
                        }
                    },
                    "204": {
                        "description": "no content"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "health checker API",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transaction/{txHash}": {
            "get": {
                "description": "get transaction information by hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Block Chain Information"
                ],
                "summary": "get transaction information by hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "txHash",
                        "name": "txHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.GetTransactionByHashRes"
                        }
                    },
                    "204": {
                        "description": "no content"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/version": {
            "get": {
                "summary": "service version API",
                "responses": {
                    "200": {
                        "description": "0.0.1",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BlockInfo": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "block_num": {
                    "type": "integer"
                },
                "block_time": {
                    "type": "integer"
                },
                "parent_hash": {
                    "type": "string"
                }
            }
        },
        "models.GetBlockByNumRes": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "block_num": {
                    "type": "integer"
                },
                "block_time": {
                    "type": "integer"
                },
                "parent_hash": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.GetLatestNBlockRes": {
            "type": "object",
            "properties": {
                "blocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BlockInfo"
                    }
                }
            }
        },
        "models.GetTransactionByHashRes": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Log"
                    }
                },
                "nonce": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Log": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                }
            }
        }
    }
}